Fix reset problem when relay turned on/off
When the relay is activated/deactivated a small volage spike is produced, the
blink led code, detect if the POWER_BUTTON is pressed with an interruption.
With the spike, the device believes that the user pressed the POWER_BUTTON,
disconnecting the device and reseting it, therefore I solved the problem by
detecting the POWER_BUTTON via software in the loop function, and now the user
needs to hold the POWER_BUTTON at least 3 seconds to disconnect the device.